07-autogpt/autogpt/auto-gpt-master/requirements.txt
ctparse==0.3.6
langchain==0.0.251
py_expression_eval==0.3.14
pydantic==1.10.12
python-dotenv==1.0.0
chromadb==0.4.5
google-search-results==2.4.2
pyowm==3.3.0
googlemaps==4.10.0